Management Web fix JsonCpp linux build setup

+++ Server/Lib/Net/NetConnectionTask.cpp
@@ -53,7 +53,7 @@ namespace Net {
 		m_pConnection = pConnection;
 		if (pConnection != nullptr)
 		{
-			SetTaskID(pConnection->GetCID());
+			SetTaskID((UINT)pConnection->GetCID());
 		}
 	}
